requirements.txt
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0

# Graph neural networks (for MAGIK)
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0

# Numerical computing
numpy>=1.24.0
scipy>=1.10.0

# Image processing
scikit-image>=0.20.0
tifffile>=2023.0.0

# Data handling
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress & logging
tqdm>=4.65.0
tensorboard>=2.12.0

# Configuration
pyyaml>=6.0

# Testing
pytest>=7.3.0

# Jupyter (optional)
jupyterlab>=4.0.0
